feat(skeleton): create in UpdatedSkeletonTree (completed)
