+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/tree.rs
@@ -7,7 +7,7 @@ use crate::patricia_merkle_tree::types::{NodeIndex, TreeHeight};
 
 use crate::storage::storage_trait::Storage;
 
-#[allow(dead_code)]
+pub(crate) type OriginalSkeletonNodeMap = HashMap<NodeIndex, OriginalSkeletonNode>;
 pub(crate) type OriginalSkeletonTreeResult<T> = Result<T, OriginalSkeletonTreeError>;
 
 /// Consider a Patricia-Merkle Tree which should be updated with new leaves.
@@ -15,7 +15,6 @@ pub(crate) type OriginalSkeletonTreeResult<T> = Result<T, OriginalSkeletonTreeEr
 /// update. It also contains the hashes (for edge siblings - also the edge data) of the Sibling
 /// nodes on the Merkle paths from the updated leaves to the root.
 pub(crate) trait OriginalSkeletonTree {
-    #[allow(dead_code)]
     fn create(
         storage: &impl Storage,
         leaf_indices: &[NodeIndex],
@@ -25,13 +24,16 @@ pub(crate) trait OriginalSkeletonTree {
     where
         Self: std::marker::Sized;
 
-    fn get_nodes(&self) -> &HashMap<NodeIndex, OriginalSkeletonNode>;
+    fn get_nodes(&self) -> &OriginalSkeletonNodeMap;
+
+    fn get_nodes_mut(&mut self) -> &mut OriginalSkeletonNodeMap;
+
+    fn get_tree_height(&self) -> &TreeHeight;
 }
 
-#[allow(dead_code)]
 #[derive(Debug, Eq, PartialEq)]
 pub(crate) struct OriginalSkeletonTreeImpl {
-    pub(crate) nodes: HashMap<NodeIndex, OriginalSkeletonNode>,
+    pub(crate) nodes: OriginalSkeletonNodeMap,
     pub(crate) tree_height: TreeHeight,
 }
 
@@ -45,7 +47,15 @@ impl OriginalSkeletonTree for OriginalSkeletonTreeImpl {
         Self::create_impl(storage, sorted_leaf_indices, root_hash, tree_height)
     }
 
-    fn get_nodes(&self) -> &HashMap<NodeIndex, OriginalSkeletonNode> {
+    fn get_nodes(&self) -> &OriginalSkeletonNodeMap {
         &self.nodes
+    }
+
+    fn get_nodes_mut(&mut self) -> &mut OriginalSkeletonNodeMap {
+        &mut self.nodes
+    }
+
+    fn get_tree_height(&self) -> &TreeHeight {
+        &self.tree_height
     }
 }
